requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
eval_type_backport==0.2.0
flet==0.25.1
flet-desktop==0.25.1
google-auth==2.36.0
griffe==1.5.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
logfire-api==2.6.2
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.57.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-ai==0.0.11
pydantic-ai-slim==0.0.11
pydantic_core==2.27.1
python-dotenv==1.0.1
pytube==15.0.0
repath==0.9.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
